requirements.txt
absl-py==0.7.1
appnope==0.1.0
astor==0.8.0
backcall==0.1.0
contextlib2==0.5.5
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.15.1
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.16.4
object-detection==0.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
tb-nightly==1.14.0a20190603
tensorflow==2.0.0b1
tensorflow-hub==0.5.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
tqdm==4.35.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
